ml-service/requirements.txt
fastapi
uvicorn
spacy
pdfplumber
python-multipart
google-generativeai
python-dotenv
